+++ src/main/scala/edu/ie3/simona/config/SimonaConfig.scala
@@ -23,9 +23,9 @@ final case class SimonaConfig(
     simona: SimonaConfig.Simona
 ) derives ConfigConvert {
 
-  /** Returns the default config values.
+  /** Returns the values of this config.
     */
-  def defaults: ConfigValue = writer.to(this)
+  def values: ConfigValue = writer.to(this)
 }
 
 object SimonaConfig {
@@ -68,19 +68,6 @@ object SimonaConfig {
     }
 
   // pure config end
-
-  /** Case class contains default and individual configs for assets.
-    * @param defaultConfig
-    *   to use
-    * @param individualConfigs
-    *   specific configs, that are used instead of the [[defaultConfig]]
-    * @tparam T
-    *   type of asset config
-    */
-  final case class AssetConfigs[T](
-      defaultConfig: T,
-      individualConfigs: List[T] = List.empty,
-  )
 
   sealed trait GridConfigParams {
     val gridIds: Option[List[String]]
@@ -125,7 +112,7 @@ object SimonaConfig {
       powerflow: Simona.Powerflow,
       runtime: RuntimeConfig = RuntimeConfig(),
       simulationName: String,
-      time: Simona.Time = Simona.Time(),
+      time: Simona.Time,
   ) derives ConfigConvert
   object Simona {
     final case class CongestionManagement(
@@ -143,7 +130,7 @@ object SimonaConfig {
     ) derives ConfigConvert
 
     final case class Powerflow(
-        maxSweepPowerDeviation: Double,
+        maxSweepPowerDeviation: Double = 1e-5,
         newtonraphson: Powerflow.Newtonraphson,
         resolution: FiniteDuration = 1.hours,
         stopOnFailure: Boolean = false,
@@ -152,14 +139,14 @@ object SimonaConfig {
     object Powerflow {
       final case class Newtonraphson(
           epsilon: List[Double] = List.empty,
-          iterations: Int,
+          iterations: Int = 50,
       ) derives ConfigConvert
     }
 
     final case class Time(
-        endDateTime: String = "2011-05-01T01:00:00Z",
+        endDateTime: String,
         schedulerReadyCheckWindow: Option[Int] = None,
-        startDateTime: String = "2011-05-01T00:00:00Z",
+        startDateTime: String,
     ) derives ConfigConvert {
       def startTime: ZonedDateTime =
         TimeUtil.withDefaults.toZonedDateTime(startDateTime)