+++ src/main/scala/edu/ie3/simona/event/listener/ResultEventListener.scala
@@ -109,6 +109,7 @@ object ResultEventListener extends Transformer3wResultSupport {
                 finalFileName,
                 new FixedResultEntityProcessor(resultClass),
                 enableCompression,
+                csv.delimiter,
               ),
             )